requirements.txt
defusedxml==0.4.1
Django==1.10.1
django-filter==0.14.0
djangorestframework==3.4.6
drf-nested-routers==0.11.1
httplib2==0.9.2
Markdown==2.6.6
oauth2==1.9.0.post1
oauthlib==2.0.0
psycopg2==2.6.2
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.21
python3-openid==3.0.10
PyYAML==3.12
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
